+++ adminapi/src/main/java/io/minio/admin/clusterinfo/DiskMetrics.java
@@ -15,30 +15,30 @@
  * limitations under the License.
  */
 
-package io.minio.admin.info;
+package io.minio.admin.clusterinfo;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Collections;
+import java.util.Map;
 
 /**
- * Objects contains the number of objects
+ * DiskMetrics has the information about XL Storage APIs
  *
- * @see <a
- *      https://github.com/minio/madmin-go/blob/main/info-commands.go#L292">info-commands.go</a>
+ * @see <a href=
+ *     "https://github.com/minio/madmin-go/blob/main/info-commands.go#L395">info-commands.go</a>
  */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Objects {
-    @JsonProperty("count")
-    private Integer count;
+public class DiskMetrics {
+  @JsonProperty("lastMinute")
+  private Map<String, TimedAction> lastMinute;
 
-    @JsonProperty("error")
-    private String error;
+  @JsonProperty("apiCalls")
+  private Map<String, String> apiCalls;
 
-    public Integer count() {
-        return count;
-    }
+  public Map<String, TimedAction> lastMinute() {
+    return Collections.unmodifiableMap(lastMinute);
+  }
 
-    public String error() {
-        return error;
-    }
+  public Map<String, String> apiCalls() {
+    return Collections.unmodifiableMap(apiCalls);
+  }
 }
